Infowindow sidelist bug fix
By design certain interaction on map can cause the infowindow to close
but the side list /clicked location is not updated. So clicking on a
location on the side list, but closing via map interaction can cause the
the same side list location not to reopen the infowindow.

Reset lastclicked if infowindow is closed via map interaction